r/aws_bedrockagent_knowledge_base: Consolidate acceptance test files.
