--- internal/service/bedrockagent/knowledge_base_managed_test.go
@@ -1,84 +1,0 @@
-// Copyright IBM Corp. 2014, 2026
-// SPDX-License-Identifier: MPL-2.0
-
-package bedrockagent_test
-
-import (
-	"fmt"
-	"testing"
-
-	awstypes "github.com/aws/aws-sdk-go-v2/service/bedrockagent/types"
-	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/statecheck"
-	"github.com/hashicorp/terraform-plugin-testing/tfjsonpath"
-	"github.com/hashicorp/terraform-provider-aws/internal/acctest"
-	tfknownvalue "github.com/hashicorp/terraform-provider-aws/internal/acctest/knownvalue"
-	"github.com/hashicorp/terraform-provider-aws/names"
-)
-
-func TestAccBedrockAgentKnowledgeBase_managed(t *testing.T) {
-	ctx := acctest.Context(t)
-	var kb awstypes.KnowledgeBase
-	rName := acctest.RandomWithPrefix(t, acctest.ResourcePrefix)
-	resourceName := "aws_bedrockagent_knowledge_base.test"
-
-	acctest.ParallelTest(ctx, t, resource.TestCase{
-		PreCheck:                 func() { acctest.PreCheck(ctx, t) },
-		ErrorCheck:               acctest.ErrorCheck(t, names.BedrockAgentServiceID),
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
-		CheckDestroy:             testAccCheckKnowledgeBaseDestroy(ctx, t),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccKnowledgeBaseConfig_managed(rName),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					testAccCheckKnowledgeBaseExists(ctx, t, resourceName, &kb),
-				),
-				ConfigStateChecks: []statecheck.StateCheck{
-					statecheck.ExpectKnownValue(resourceName, tfjsonpath.New("knowledge_base_configuration").AtSliceIndex(0).AtMapKey(names.AttrType), tfknownvalue.StringExact(awstypes.KnowledgeBaseTypeManaged)),
-				},
-			},
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
-
-func testAccKnowledgeBaseConfig_managed(rName string) string {
-	return fmt.Sprintf(`
-data "aws_caller_identity" "current" {}
-
-resource "aws_iam_role" "test" {
-  name = %[1]q
-
-  assume_role_policy = jsonencode({
-    Version = "2012-10-17"
-    Statement = [{
-      Effect    = "Allow"
-      Principal = { Service = "bedrock.amazonaws.com" }
-      Action    = "sts:AssumeRole"
-      Condition = {
-        StringEquals = {
-          "aws:SourceAccount" = data.aws_caller_identity.current.account_id
-        }
-      }
-    }]
-  })
-}
-
-resource "aws_bedrockagent_knowledge_base" "test" {
-  name     = %[1]q
-  role_arn = aws_iam_role.test.arn
-
-  knowledge_base_configuration {
-    type = "MANAGED"
-
-    managed_knowledge_base_configuration {
-      embedding_model_type = "MANAGED"
-    }
-  }
-}
-`, rName)
-}
